api/requirements.txt
alembic==1.13.1
asyncpg==0.29.0
fastapi==0.111.0
httpx[socks]==0.27.0
passlib==1.7.4
pycryptodome==3.20.0
pydantic==2.7.2
python-jose==3.3.0
python-multipart==0.0.9
sqlalchemy[asyncio]==2.0.30
user-agents==2.2.0
uvicorn==0.30.0
boto3==1.34.115
python-jose==3.3.0